Add the client and server middleware
To make the handling of GlobalID-compatible objects truly transparent,
we need to handle the logic within Sidekiq middleware. Middleware make
it so you cannot forget to include a module to handle the serialization
or remember the scheme that a worker uses to look up an object.

This iteration of the middleware handles everything except for Hashes
with keys or values that are GlobalID-compatible.