Address refinement differences between 3.0 and 2.7
A (seemingly?) undocumented feature in Ruby 3.0 is that you can refine a
module and have the refinement apply to including classes, e.g. refining
`Enumerable` allows the refinement to apply to `Array`. However, this is
not the case in 2.7 so we need to muddy up the refinement to account for
that on older versions of Ruby.

This likely means that we will need to handle ActiveRecord relations
separately since they will not acquire the serialization refinements via
`Enumerable`.