Add Java platforms to Gemfile.lock
This is something I'm not used to. I've recently come to reconsider
including a Gemfile.lock in the repository to help new contributors
onboard to the project. As such, there's a bit more song-and-dance to
handle JRuby in CI, like adding platform information to the Gemfile.lock
file.

It might make sense to let CI install the latest versions of gems
instead, but I'm going to try this approach for now.