Add support for ActiveRecord::Relations
On Ruby v3.0+, refinements work on module methods. This means that
anything that includes `Enumerable` can be refined by refining
`Enumerable`. However, on older Rubies, this is not the case.

Since people often use Sidekiq within Rails applications, we need to
support ActiveRecord relations within the middleware. This necessitated
a complicated integration test that runs in a separate process so that
the refinement is able to detect the presence of ActiveRecord::Relation
when writing itself.

The reason for this is that you cannot override the lexical binding of
refinements by later extending the refinement. Thus, the refinement
definition _must_ come after Rails has loaded ActiveRecord. This will be
the default behavior in any Rails application since the Rails templates
load their own gems before any non-Rails gems.

This might introduce some maintenance burden for people with advanced
Rails setups, but I don't think we should optimize for that until it's
reported as a problem. Adding it to the readme would only confuse people
who use normal Rails conventions.